+++ src/shuriken/monkey_patch.clj
@@ -327,7 +327,9 @@
   (let [ns-str (-> namespaced-sym .getNamespace)
         _ (assert ns-str "Symbol must be namespaced")
         ns-sym (symbol ns-str)
-        source (read-string (source-fn namespaced-sym))
+        source-str (source-fn namespaced-sym)
+        _ (assert source-str (format "No source found for %s" namespaced-sym))
+        source (read-string source-str)
         _ (assert source
                   (str "No source found for symbol '" namespaced-sym "'"))
         v (resolve namespaced-sym)]